% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunction.R
\name{myFunction}
\alias{myFunction}
\title{Computes means by groups}
\usage{
myFunction(values, groups)
}
\arguments{
\item{values}{numeric vector of values}

\item{groups}{vector of groups}
}
\value{
data.frame with groups names in first column and average values per group in the second one
}
\description{
You can put extended description here
}
\details{
If your function is complicated, you may consider adding a details section
}

